Fixed calculation goods and discounts implemented

+++ api/src/main/java/terabu/dto/bucket/BucketResponse.java
@@ -9,5 +9,5 @@ public class BucketResponse {
     private Long id;
     private String nameGoods;
     private Long count;
-    private Long sum;
+    private Double sum;
 }
